transformacion/dbt/requirements.txt
dbt-core==1.9.4 
dbt-athena-community==1.9.4
dbt-duckdb==1.9.4
shandy-sqlfmt[jinjafmt]==0.26.0